Added clock component to home page
A new clock component has been added to the home page. This component displays the current time, updated every second. The clock is styled with a large monospace font and centered on the page. The home page layout has been adjusted to center this new component, and its background color has been changed to black for better contrast with the white text of the clock.