Wire up the time offset from the settings page
The commit introduces the 'date-fns' library to the project dependencies. This addition is used in implementing a new feature that allows for an offset time display on the clock component. The clock component now uses settings from the SettingsService to calculate and display an offset time, instead of just displaying the current time.