Added scrolling message feature
A new package 'ngx-fast-marquee' has been added to the project dependencies. This package is used to implement a scrolling message feature on the home page. The HTML, SCSS, and TypeScript files of the ScrollingMessageComponent have been updated accordingly. The width of the app-scrolling-message in MessagePageComponent's SCSS file has also been set to 100%.