Added screen lock status to home page
In this update, we've introduced a new feature that displays the status of the screen lock on the home page. This is represented by two icons: one for when the screen lock is enabled and another for when it fails. We've also added a new service called WakeLockService which handles the logic behind enabling or disabling the screen lock.